+++ src/anime_parsers_ru/parser_aniboom_async.py
@@ -232,6 +232,8 @@ class AniboomParserAsync:
         """
         c_data = {}
         response = await self.requests.get(link)
+        if response.status_code == 429:
+            raise errors.TooManyRequests('Сервер вернул код ошибки 429. Слишком частые запросы')
         if response.status_code != 200:
             raise errors.ServiceError(f'Сервер не вернул ожидаемый код 200. Код: "{response.status_code}"')
         response = response.text
@@ -308,6 +310,8 @@ class AniboomParserAsync:
             '_allow': 'true',
         }
         response = await self.requests.get(f'https://{self._dmn}/anime/{animego_id}/player', params=params, headers=headers)
+        if response.status_code == 429:
+            raise errors.TooManyRequests('Сервер вернул код ошибки 429. Слишком частые запросы')
         if response.status_code != 200:
             raise errors.ServiceError(f'Сервер не вернул ожидаемый код 200. Код: "{response.status_code}"')
         response = response.json()
@@ -356,6 +360,8 @@ class AniboomParserAsync:
             '_allow': 'true',
         }
         response = await self.requests.get(f'https://{self._dmn}/anime/{animego_id}/player', params=params, headers=headers)
+        if response.status_code == 429:
+            raise errors.TooManyRequests('Сервер вернул код ошибки 429. Слишком частые запросы')
         if response.status_code != 200:
             raise errors.ServiceError(f'Сервер не вернул ожидаемый код 200. Код: "{response.status_code}"')
         response = response.json()
@@ -394,6 +400,8 @@ class AniboomParserAsync:
                 'translation': translation,
             }
         response = await self.requests.get(embed_link, headers=headers, params=params)
+        if response.status_code == 429:
+            raise errors.TooManyRequests('Сервер вернул код ошибки 429. Слишком частые запросы')
         if response.status_code != 200:
             raise errors.ServiceError(f'Сервер не вернул ожидаемый код 200. Код: "{response.status_code}"')
         return response.text
@@ -461,8 +469,13 @@ class AniboomParserAsync:
         playlist = await self.requests.get(media_src, headers=headers)
         playlist = playlist.text
         # Вставляем полный путь до сервера
-        server_path = media_src[:media_src.rfind('master_device.m3u8')]
-        playlist = playlist.replace('media_', server_path+'media_')
+        if '<MPD' in playlist:
+            filename = media_src[media_src.rfind('/')+1:media_src.rfind('.')]
+            server_path = media_src[:media_src.rfind('.')]
+            playlist = playlist.replace(filename, server_path)
+        else:
+            server_path = media_src[:media_src.rfind('master_device.m3u8')]
+            playlist = playlist.replace('media_', server_path+'media_')
         return playlist
     
     async def get_mpd_playlist(self, animego_id: str, episode: int, translation_id: str) -> str: